form layout change

+++ database/migrations/2023_11_04_064507_add_colomns_in_exams_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::table(config('table.exams'), function (Blueprint $table) {
             $table->after('id', function (Blueprint $table) {
                 $table->string('name');
-                $table->unsignedTinyInteger('subject_id');
-                $table->unsignedTinyInteger('class_room_id');
+                $table->unsignedSmallInteger('subject_id');
+                $table->unsignedSmallInteger('class_room_id');
 
             });
         });
